Struct scheduled_thread_pool::builder::FinalStage

pub struct FinalStage<'a> { /* private fields */ }

The final builder stage, allowing configuration of optional paramters.

Implementations

impl<'a> FinalStage<'a>

pub fn thread_name_pattern(self, thread_name_pattern: &'a str) -> Self

Sets the pattern to be used when naming threads created to be part of the pool.

The substring {} in the name will be replaced with an integer identifier of the thread.

Defaults to None.

pub fn on_drop_behavior(self, on_drop_behavior: OnPoolDropBehavior) -> Self

Sets the behavior for what to do with pending scheduled executions when the pool is dropped.

Defaults to OnPoolDropBehavior::CompletePendingScheduled.

pub fn build(self) -> ScheduledThreadPool

Consumes the builder, returning the constructed thread pool.