Crate scheduled_thread_pool

A thread pool to execute scheduled actions in parallel.

While a normal thread pool is only able to execute actions as soon as possible, a scheduled thread pool can execute actions after a specific delay, or execute actions periodically.

Modules

builder

Staged builder types.

Structs

JobHandle

A handle to a scheduled job.

ScheduledThreadPool

A pool of threads which can run tasks at specific time intervals.

Enums

OnPoolDropBehavior

Options for what the behavior should be in regards to pending scheduled executions when the pool is dropped.